# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "renn-ai-crm API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "Login with email/password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/webhook": {
      "post": {
        "summary": "Webhook receiver (HMAC + optional replay guard)",
        "parameters": [
          {
            "in": "header",
            "name": "x-id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-timestamp",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-signature",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid"
          },
          "401": {
            "description": "Missing signature"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "LoginBody": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 256
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 256
          }
        }
      },
...
    }
  }
}